36-separate-request-processing: BROKEN WIP
